requirements.txt
requests
pymodbus
python-snap7
fastapi
uvicorn
python-multipart
jinja2
